code cleanup
-massive code cleanup
-created abstract Switcher class
-switcherX and switcherC extend Switcher class
