+++ library/src/main/java/com/bitvale/switcher/SwitcherC.kt
@@ -3,20 +3,13 @@ package com.bitvale.switcher
 import android.animation.AnimatorSet
 import android.animation.ArgbEvaluator
 import android.animation.ValueAnimator
-import android.annotation.SuppressLint
 import android.annotation.TargetApi
 import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.Color
 import android.graphics.Outline
-import android.graphics.Paint
-import android.graphics.PorterDuff
-import android.graphics.PorterDuffColorFilter
-import android.graphics.RectF
 import android.os.Build
-import android.os.Bundle
-import android.os.Parcelable
 import android.renderscript.Allocation
 import android.renderscript.Element
 import android.renderscript.RenderScript
@@ -24,19 +17,8 @@ import android.renderscript.ScriptIntrinsicBlur
 import android.util.AttributeSet
 import android.view.View
 import android.view.ViewOutlineProvider
-import androidx.annotation.ColorInt
 import androidx.core.animation.doOnStart
-import com.bitvale.switcher.commons.BOUNCE_ANIM_AMPLITUDE_IN
-import com.bitvale.switcher.commons.BOUNCE_ANIM_AMPLITUDE_OUT
-import com.bitvale.switcher.commons.BOUNCE_ANIM_FREQUENCY_IN
-import com.bitvale.switcher.commons.BOUNCE_ANIM_FREQUENCY_OUT
-import com.bitvale.switcher.commons.COLOR_ANIMATION_DURATION
-import com.bitvale.switcher.commons.KEY_CHECKED
-import com.bitvale.switcher.commons.STATE
-import com.bitvale.switcher.commons.SWITCHER_ANIMATION_DURATION
-import com.bitvale.switcher.commons.isLollipopAndAbove
-import com.bitvale.switcher.commons.lerp
-import com.bitvale.switcher.commons.toPx
+import com.bitvale.switcher.commons.*
 import kotlin.math.min
 
 
@@ -47,112 +29,9 @@ class SwitcherC @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0
-) : View(context, attrs, defStyleAttr) {
+) : Switcher(context, attrs, defStyleAttr) {
 
     private var switcherRadius = 0f
-    private var iconRadius = 0f
-    private var iconClipRadius = 0f
-    private var iconCollapsedWidth = 0f
-    private var defHeight = 0
-    private var defWidth = 0
-    var isChecked = true
-        private set
-
-    @ColorInt
-    var onColor = 0
-    @ColorInt
-    var offColor = 0
-    @ColorInt
-    private var iconColor = 0
-
-    private val switcherPaint = Paint(Paint.ANTI_ALIAS_FLAG)
-
-    private val iconRect = RectF(0f, 0f, 0f, 0f)
-    private val iconClipRect = RectF(0f, 0f, 0f, 0f)
-    private val iconPaint = Paint(Paint.ANTI_ALIAS_FLAG)
-
-    private val iconClipPaint = Paint(Paint.ANTI_ALIAS_FLAG)
-
-    private var animatorSet: AnimatorSet? = AnimatorSet()
-
-    private val shadowPaint = Paint(Paint.ANTI_ALIAS_FLAG)
-    private var shadow: Bitmap? = null
-    private var shadowOffset = 0f
-
-    @ColorInt
-    private var currentColor = 0
-        set(value) {
-            field = value
-            switcherPaint.color = value
-            iconClipPaint.color = value
-        }
-
-    private var switchElevation = 0f
-    private var iconHeight = 0f
-
-    // from rounded rect to circle and back
-    private var iconProgress = 0f
-        set(value) {
-            if (field != value) {
-                field = value
-
-                val iconOffset = lerp(0f, iconRadius - iconCollapsedWidth / 2, value)
-                iconRect.left =
-                    (switcherRadius - iconCollapsedWidth / 2 - iconOffset) + shadowOffset
-                iconRect.right =
-                    (switcherRadius + iconCollapsedWidth / 2 + iconOffset) + shadowOffset
-
-                val clipOffset = lerp(0f, iconClipRadius, value)
-                iconClipRect.set(
-                    iconRect.centerX() - clipOffset,
-                    iconRect.centerY() - clipOffset,
-                    iconRect.centerX() + clipOffset,
-                    iconRect.centerY() + clipOffset
-                )
-                postInvalidateOnAnimation()
-            }
-        }
-
-    init {
-        attrs?.let { retrieveAttributes(attrs, defStyleAttr) }
-        setOnClickListener { setChecked(!isChecked) }
-    }
-
-    @SuppressLint("CustomViewStyleable")
-    private fun retrieveAttributes(attrs: AttributeSet, defStyleAttr: Int) {
-        val typedArray = context.obtainStyledAttributes(
-            attrs,
-            R.styleable.Switcher,
-            defStyleAttr,
-            R.style.Switcher
-        )
-
-        switchElevation = typedArray.getDimension(R.styleable.Switcher_elevation, 0f)
-
-        onColor = typedArray.getColor(R.styleable.Switcher_switcher_on_color, 0)
-        offColor = typedArray.getColor(R.styleable.Switcher_switcher_off_color, 0)
-        iconColor = typedArray.getColor(R.styleable.Switcher_switcher_icon_color, 0)
-
-        isChecked = typedArray.getBoolean(R.styleable.Switcher_android_checked, true)
-
-        if (!isChecked) iconProgress = 1f
-
-        currentColor = if (isChecked) onColor else offColor
-
-        iconPaint.color = iconColor
-
-        defHeight = typedArray.getDimensionPixelOffset(R.styleable.Switcher_switcher_height, 0)
-        defWidth = typedArray.getDimensionPixelOffset(R.styleable.Switcher_switcher_width, 0)
-
-        typedArray.recycle()
-
-        if (!isLollipopAndAbove() && switchElevation > 0f) {
-            shadowPaint.colorFilter = PorterDuffColorFilter(Color.BLACK, PorterDuff.Mode.SRC_IN)
-            shadowPaint.alpha = 51 // 20%
-            setShadowBlurRadius(switchElevation)
-            setLayerType(LAYER_TYPE_SOFTWARE, null)
-        }
-    }
 
     override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
         val widthMode = MeasureSpec.getMode(widthMeasureSpec)
@@ -175,8 +54,6 @@ class SwitcherC @JvmOverloads constructor(
     }
 
     override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
-        super.onSizeChanged(w, h, oldw, oldh)
-
         if (isLollipopAndAbove()) {
             outlineProvider = SwitchOutline(w, h)
             elevation = switchElevation
@@ -193,30 +70,30 @@ class SwitcherC @JvmOverloads constructor(
         iconHeight = iconRadius * 2f
 
         iconRect.set(
-            (switcherRadius - iconCollapsedWidth / 2f) + shadowOffset,
-            ((switcherRadius * 2f - iconHeight) / 2f) + shadowOffset / 2,
-            (switcherRadius + iconCollapsedWidth / 2f) + shadowOffset,
-            (switcherRadius * 2f - (switcherRadius * 2f - iconHeight) / 2f) + shadowOffset / 2
+                (switcherRadius - iconCollapsedWidth / 2f) + shadowOffset,
+                ((switcherRadius * 2f - iconHeight) / 2f) + shadowOffset / 2,
+                (switcherRadius + iconCollapsedWidth / 2f) + shadowOffset,
+                (switcherRadius * 2f - (switcherRadius * 2f - iconHeight) / 2f) + shadowOffset / 2
         )
 
         if (!isChecked) {
             iconRect.left =
-                (switcherRadius - iconCollapsedWidth / 2f - (iconRadius - iconCollapsedWidth / 2f)) + shadowOffset
+                    (switcherRadius - iconCollapsedWidth / 2f - (iconRadius - iconCollapsedWidth / 2f)) + shadowOffset
             iconRect.right =
-                (switcherRadius + iconCollapsedWidth / 2f + (iconRadius - iconCollapsedWidth / 2f)) + shadowOffset
+                    (switcherRadius + iconCollapsedWidth / 2f + (iconRadius - iconCollapsedWidth / 2f)) + shadowOffset
 
             iconClipRect.set(
-                iconRect.centerX() - iconClipRadius,
-                iconRect.centerY() - iconClipRadius,
-                iconRect.centerX() + iconClipRadius,
-                iconRect.centerY() + iconClipRadius
+                    iconRect.centerX() - iconClipRadius,
+                    iconRect.centerY() - iconClipRadius,
+                    iconRect.centerX() + iconClipRadius,
+                    iconRect.centerY() + iconClipRadius
             )
         }
 
         if (!isLollipopAndAbove()) generateShadow()
     }
 
-    private fun generateShadow() {
+    override fun generateShadow() {
         if (switchElevation == 0f) return
         if (!isInEditMode) {
             if (shadow == null) {
@@ -227,8 +104,8 @@ class SwitcherC @JvmOverloads constructor(
             val c = Canvas(shadow as Bitmap)
 
             c.drawCircle(
-                switcherRadius + shadowOffset, switcherRadius + shadowOffset / 2,
-                switcherRadius, shadowPaint
+                    switcherRadius + shadowOffset, switcherRadius + shadowOffset / 2,
+                    switcherRadius, shadowPaint
             )
             val rs = RenderScript.create(context)
             val blur = ScriptIntrinsicBlur.create(rs, Element.U8(rs))
@@ -244,17 +121,6 @@ class SwitcherC @JvmOverloads constructor(
         }
     }
 
-    fun applyColorChange() {
-        animatorSet?.cancel()
-        if (!isChecked) {
-            currentColor = offColor
-            iconProgress = 1f
-        } else {
-            currentColor = onColor
-            iconProgress = 0f
-        }
-    }
-
     override fun onDraw(canvas: Canvas?) {
         // shadow
         if (!isLollipopAndAbove() && switchElevation > 0f && !isInEditMode) {
@@ -263,8 +129,8 @@ class SwitcherC @JvmOverloads constructor(
 
         // switcher
         canvas?.drawCircle(
-            switcherRadius + shadowOffset, switcherRadius + shadowOffset / 2,
-            switcherRadius, switcherPaint
+                switcherRadius + shadowOffset, switcherRadius + shadowOffset / 2,
+                switcherRadius, switcherPaint
         )
 
         // icon
@@ -276,7 +142,7 @@ class SwitcherC @JvmOverloads constructor(
 
     }
 
-    private fun animateSwitch() {
+    override fun animateSwitch() {
         animatorSet?.cancel()
         animatorSet = AnimatorSet()
 
@@ -318,25 +184,7 @@ class SwitcherC @JvmOverloads constructor(
         }
     }
 
-    private var listener: ((isChecked: Boolean) -> Unit)? = null
-
-    /**
-     * Register a callback to be invoked when the isChecked state of this switch
-     * changes.
-     *
-     * @param listener the callback to call on isChecked state change
-     */
-    fun setOnCheckedChangeListener(listener: (isChecked: Boolean) -> Unit) {
-        this.listener = listener
-    }
-
-    /**
-     * <p>Changes the isChecked state of this switch.</p>
-     *
-     * @param checked true to check the switch, false to uncheck it
-     * @param withAnimation use animation
-     */
-    fun setChecked(checked: Boolean, withAnimation: Boolean = true) {
+    override fun setChecked(checked: Boolean, withAnimation: Boolean) {
         if (this.isChecked != checked) {
             this.isChecked = checked
             if (withAnimation && width != 0) {
@@ -354,47 +202,27 @@ class SwitcherC @JvmOverloads constructor(
         }
     }
 
-    override fun onSaveInstanceState(): Parcelable {
-        super.onSaveInstanceState()
-        return Bundle().apply {
-            putBoolean(KEY_CHECKED, isChecked)
-            putParcelable(STATE, super.onSaveInstanceState())
-        }
-    }
-
-    override fun onRestoreInstanceState(state: Parcelable?) {
-        if (state is Bundle) {
-            super.onRestoreInstanceState(state.getParcelable(STATE))
-            isChecked = state.getBoolean(KEY_CHECKED)
-            if (!isChecked) forceUncheck()
-        }
-    }
-
-    private fun forceUncheck() {
+    override fun forceUncheck() {
         currentColor = offColor
         iconProgress = 1f
     }
 
-    private fun setShadowBlurRadius(elevation: Float) {
-        val maxElevation = context.toPx(24f)
-        switchElevation = min(25f * (elevation / maxElevation), 25f)
-    }
-
     @TargetApi(Build.VERSION_CODES.LOLLIPOP)
     private inner class SwitchOutline internal constructor(
-        internal var width: Int,
-        internal var height: Int
+            internal var width: Int,
+            internal var height: Int
     ) :
-        ViewOutlineProvider() {
+            ViewOutlineProvider() {
 
         override fun getOutline(view: View, outline: Outline) {
             outline.setRoundRect(
-                0,
-                0,
-                (switcherRadius * 2).toInt(),
-                (switcherRadius * 2).toInt(),
-                switcherRadius
+                    0,
+                    0,
+                    (switcherRadius * 2).toInt(),
+                    (switcherRadius * 2).toInt(),
+                    switcherRadius
             )
         }
     }
+
 }
