+++ library/src/main/java/com/bitvale/switcher/SwitcherX.kt
@@ -3,20 +3,10 @@ package com.bitvale.switcher
 import android.animation.AnimatorSet
 import android.animation.ArgbEvaluator
 import android.animation.ValueAnimator
-import android.annotation.SuppressLint
 import android.annotation.TargetApi
 import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.Color
-import android.graphics.Outline
-import android.graphics.Paint
-import android.graphics.PorterDuff
-import android.graphics.PorterDuffColorFilter
-import android.graphics.RectF
+import android.graphics.*
 import android.os.Build
-import android.os.Bundle
-import android.os.Parcelable
 import android.renderscript.Allocation
 import android.renderscript.Element
 import android.renderscript.RenderScript
@@ -24,151 +14,26 @@ import android.renderscript.ScriptIntrinsicBlur
 import android.util.AttributeSet
 import android.view.View
 import android.view.ViewOutlineProvider
-import androidx.annotation.ColorInt
 import androidx.core.animation.doOnEnd
 import androidx.core.animation.doOnStart
 import androidx.core.graphics.withTranslation
-import com.bitvale.switcher.commons.BOUNCE_ANIM_AMPLITUDE_IN
-import com.bitvale.switcher.commons.BOUNCE_ANIM_AMPLITUDE_OUT
-import com.bitvale.switcher.commons.BOUNCE_ANIM_FREQUENCY_IN
-import com.bitvale.switcher.commons.BOUNCE_ANIM_FREQUENCY_OUT
-import com.bitvale.switcher.commons.COLOR_ANIMATION_DURATION
-import com.bitvale.switcher.commons.KEY_CHECKED
-import com.bitvale.switcher.commons.ON_CLICK_RADIUS_OFFSET
-import com.bitvale.switcher.commons.STATE
-import com.bitvale.switcher.commons.SWITCHER_ANIMATION_DURATION
-import com.bitvale.switcher.commons.TRANSLATE_ANIMATION_DURATION
-import com.bitvale.switcher.commons.isLollipopAndAbove
-import com.bitvale.switcher.commons.lerp
-import com.bitvale.switcher.commons.toPx
-import kotlin.math.min
+import com.bitvale.switcher.commons.*
 
 /**
  * Created by Alexander Kolpakov on 11/7/2018
  */
 class SwitcherX @JvmOverloads constructor(
-    context: Context,
-    attrs: AttributeSet? = null,
-    defStyleAttr: Int = 0
-) : View(context, attrs, defStyleAttr) {
-
-    private var iconRadius = 0f
-    private var iconClipRadius = 0f
-    private var iconCollapsedWidth = 0f
-    private var defHeight = 0
-    private var defWidth = 0
-    var isChecked = true
-        private set
-
-    @ColorInt
-    var onColor = 0
-    @ColorInt
-    var offColor = 0
-    @ColorInt
-    private var iconColor = 0
+        context: Context,
+        attrs: AttributeSet? = null,
+        defStyleAttr: Int = 0
+) : Switcher(context, attrs, defStyleAttr) {
 
     private val switcherRect = RectF(0f, 0f, 0f, 0f)
-    private val switcherPaint = Paint(Paint.ANTI_ALIAS_FLAG)
-
-    private val iconRect = RectF(0f, 0f, 0f, 0f)
-    private val iconClipRect = RectF(0f, 0f, 0f, 0f)
-    private val iconPaint = Paint(Paint.ANTI_ALIAS_FLAG)
-
-    private val iconClipPaint = Paint(Paint.ANTI_ALIAS_FLAG)
-
-    private val shadowPaint = Paint(Paint.ANTI_ALIAS_FLAG)
-    private var shadow: Bitmap? = null
-    private var shadowOffset = 0f
-
-    private var animatorSet: AnimatorSet? = AnimatorSet()
-
-    private var onClickOffset = 0f
-        set(value) {
-            field = value
-            switcherRect.left = value + shadowOffset
-            switcherRect.top = value + shadowOffset / 2
-            switcherRect.right = width.toFloat() - value - shadowOffset
-            switcherRect.bottom = height.toFloat() - value - shadowOffset - shadowOffset / 2
-            if (!isLollipopAndAbove()) generateShadow()
-            invalidate()
-        }
-
-    @ColorInt
-    private var currentColor = 0
-        set(value) {
-            field = value
-            switcherPaint.color = value
-            iconClipPaint.color = value
-        }
-
     private var switcherCornerRadius = 0f
-    private var switchElevation = 0f
-    private var iconHeight = 0f
 
     private var iconTranslateX = 0f
 
-    // from rounded rect to circle and back
-    private var iconProgress = 0f
-        set(value) {
-            if (field != value) {
-                field = value
-
-                val iconOffset = lerp(0f, iconRadius - iconCollapsedWidth / 2, value)
-                iconRect.left = width - switcherCornerRadius - iconCollapsedWidth / 2 - iconOffset
-                iconRect.right = width - switcherCornerRadius + iconCollapsedWidth / 2 + iconOffset
-
-                val clipOffset = lerp(0f, iconClipRadius, value)
-                iconClipRect.set(
-                    iconRect.centerX() - clipOffset,
-                    iconRect.centerY() - clipOffset,
-                    iconRect.centerX() + clipOffset,
-                    iconRect.centerY() + clipOffset
-                )
-                if (!isLollipopAndAbove()) generateShadow()
-                postInvalidateOnAnimation()
-            }
-        }
-
-    init {
-        attrs?.let { retrieveAttributes(attrs, defStyleAttr) }
-        setOnClickListener { setChecked(!isChecked) }
-    }
-
-    @SuppressLint("CustomViewStyleable")
-    private fun retrieveAttributes(attrs: AttributeSet, defStyleAttr: Int) {
-        val typedArray = context.obtainStyledAttributes(
-            attrs,
-            R.styleable.Switcher,
-            defStyleAttr,
-            R.style.Switcher
-        )
-
-        switchElevation = typedArray.getDimension(R.styleable.Switcher_elevation, 0f)
-
-        onColor = typedArray.getColor(R.styleable.Switcher_switcher_on_color, 0)
-        offColor = typedArray.getColor(R.styleable.Switcher_switcher_off_color, 0)
-        iconColor = typedArray.getColor(R.styleable.Switcher_switcher_icon_color, 0)
-
-        isChecked = typedArray.getBoolean(R.styleable.Switcher_android_checked, true)
-
-        if (!isChecked) iconProgress = 1f
-
-        currentColor = if (isChecked) onColor else offColor
-
-        iconPaint.color = iconColor
-
-        defHeight = typedArray.getDimensionPixelOffset(R.styleable.Switcher_switcher_height, 0)
-        defWidth = typedArray.getDimensionPixelOffset(R.styleable.Switcher_switcher_width, 0)
-
-        typedArray.recycle()
-
-        if (!isLollipopAndAbove() && switchElevation > 0f) {
-            shadowPaint.colorFilter = PorterDuffColorFilter(Color.BLACK, PorterDuff.Mode.SRC_IN)
-            shadowPaint.alpha = 51 // 20%
-            setShadowBlurRadius(switchElevation)
-            setLayerType(LAYER_TYPE_SOFTWARE, null)
-        }
-    }
+    private var onClickOffset = 0f
 
     override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
         val widthMode = MeasureSpec.getMode(widthMeasureSpec)
@@ -190,7 +55,6 @@ class SwitcherX @JvmOverloads constructor(
     }
 
     override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
-        super.onSizeChanged(w, h, oldw, oldh)
 
         if (isLollipopAndAbove()) {
             outlineProvider = SwitchOutline(w, h)
@@ -214,23 +78,23 @@ class SwitcherX @JvmOverloads constructor(
         iconHeight = iconRadius * 2f
 
         iconRect.set(
-            width - switcherCornerRadius - iconCollapsedWidth / 2,
-            ((height - iconHeight) / 2f) - shadowOffset / 2,
-            width - switcherCornerRadius + iconCollapsedWidth / 2,
-            (height - (height - iconHeight) / 2f) - shadowOffset / 2
+                width - switcherCornerRadius - iconCollapsedWidth / 2,
+                ((height - iconHeight) / 2f) - shadowOffset / 2,
+                width - switcherCornerRadius + iconCollapsedWidth / 2,
+                (height - (height - iconHeight) / 2f) - shadowOffset / 2
         )
 
         if (!isChecked) {
             iconRect.left =
-                width - switcherCornerRadius - iconCollapsedWidth / 2 - (iconRadius - iconCollapsedWidth / 2)
+                    width - switcherCornerRadius - iconCollapsedWidth / 2 - (iconRadius - iconCollapsedWidth / 2)
             iconRect.right =
-                width - switcherCornerRadius + iconCollapsedWidth / 2 + (iconRadius - iconCollapsedWidth / 2)
+                    width - switcherCornerRadius + iconCollapsedWidth / 2 + (iconRadius - iconCollapsedWidth / 2)
 
             iconClipRect.set(
-                iconRect.centerX() - iconClipRadius,
-                iconRect.centerY() - iconClipRadius,
-                iconRect.centerX() + iconClipRadius,
-                iconRect.centerY() + iconClipRadius
+                    iconRect.centerX() - iconClipRadius,
+                    iconRect.centerY() - iconClipRadius,
+                    iconRect.centerX() + iconClipRadius,
+                    iconRect.centerY() + iconClipRadius
             )
 
             iconTranslateX = -(width - shadowOffset - switcherCornerRadius * 2)
@@ -239,7 +103,7 @@ class SwitcherX @JvmOverloads constructor(
         if (!isLollipopAndAbove()) generateShadow()
     }
 
-    private fun generateShadow() {
+    override fun generateShadow() {
         if (switchElevation == 0f) return
         if (!isInEditMode) {
             if (shadow == null) {
@@ -265,17 +129,6 @@ class SwitcherX @JvmOverloads constructor(
         }
     }
 
-    fun applyColorChange() {
-        animatorSet?.cancel()
-        if (!isChecked) {
-            currentColor = offColor
-            iconProgress = 1f
-        } else {
-            currentColor = onColor
-            iconProgress = 0f
-        }
-    }
-
     override fun onDraw(canvas: Canvas?) {
         // shadow
         if (!isLollipopAndAbove() && switchElevation > 0f && !isInEditMode) {
@@ -284,15 +137,15 @@ class SwitcherX @JvmOverloads constructor(
 
         // switcher
         canvas?.drawRoundRect(
-            switcherRect,
-            switcherCornerRadius,
-            switcherCornerRadius,
-            switcherPaint
+                switcherRect,
+                switcherCornerRadius,
+                switcherCornerRadius,
+                switcherPaint
         )
 
         // icon
         canvas?.withTranslation(
-            x = iconTranslateX
+                x = iconTranslateX
         ) {
             drawRoundRect(iconRect, switcherCornerRadius, switcherCornerRadius, iconPaint)
             /* don't draw clip path if icon is collapsed (to prevent drawing small circle
@@ -302,7 +155,7 @@ class SwitcherX @JvmOverloads constructor(
         }
     }
 
-    private fun animateSwitch() {
+    override fun animateSwitch() {
         animatorSet?.cancel()
         animatorSet = AnimatorSet()
 
@@ -359,25 +212,7 @@ class SwitcherX @JvmOverloads constructor(
         }
     }
 
-    private var listener: ((isChecked: Boolean) -> Unit)? = null
-
-    /**
-     * Register a callback to be invoked when the isChecked state of this switch
-     * changes.
-     *
-     * @param listener the callback to call on isChecked state change
-     */
-    fun setOnCheckedChangeListener(listener: (isChecked: Boolean) -> Unit) {
-        this.listener = listener
-    }
-
-    /**
-     * <p>Changes the isChecked state of this switch.</p>
-     *
-     * @param checked true to check the switch, false to uncheck it
-     * @param withAnimation use animation
-     */
-    fun setChecked(checked: Boolean, withAnimation: Boolean = true) {
+    override fun setChecked(checked: Boolean, withAnimation: Boolean) {
         if (this.isChecked != checked) {
             this.isChecked = checked
             if (withAnimation && width != 0) {
@@ -397,42 +232,22 @@ class SwitcherX @JvmOverloads constructor(
         }
     }
 
-    override fun onSaveInstanceState(): Parcelable {
-        super.onSaveInstanceState()
-        return Bundle().apply {
-            putBoolean(KEY_CHECKED, isChecked)
-            putParcelable(STATE, super.onSaveInstanceState())
-        }
-    }
-
-    override fun onRestoreInstanceState(state: Parcelable?) {
-        if (state is Bundle) {
-            super.onRestoreInstanceState(state.getParcelable(STATE))
-            isChecked = state.getBoolean(KEY_CHECKED)
-            if (!isChecked) forceUncheck()
-        }
-    }
-
-    private fun forceUncheck() {
+    override fun forceUncheck() {
         currentColor = offColor
         iconProgress = 1f
         iconTranslateX = -(width - shadowOffset - switcherCornerRadius * 2)
     }
 
-    private fun setShadowBlurRadius(elevation: Float) {
-        val maxElevation = context.toPx(24f)
-        switchElevation = min(25f * (elevation / maxElevation), 25f)
-    }
-
     @TargetApi(Build.VERSION_CODES.LOLLIPOP)
     private inner class SwitchOutline internal constructor(
-        internal var width: Int,
-        internal var height: Int
+            internal var width: Int,
+            internal var height: Int
     ) :
-        ViewOutlineProvider() {
+            ViewOutlineProvider() {
 
         override fun getOutline(view: View, outline: Outline) {
             outline.setRoundRect(0, 0, width, height, switcherCornerRadius)
         }
     }
+
 }
